requirements.txt
beautifulsoup4==4.9.3
chromedriver-binary-auto==0.1
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
selenium==3.141.0
